錄音轉文字工具/requirements.txt
streamlit
faster-whisper
opencc-python-reimplemented
